Now build system works!

+++ main.c
@@ -21,7 +21,9 @@ enum ParseMode{
 	CURPARSE_FILES,
 	CURPARSE_COMPILERFLAGS,
 	CURPARSE_ASSEMBLERFLAGS,
-	CURPARSE_LINKERFLAGS
+	CURPARSE_LINKERFLAGS,
+	CURPARSE_OUTPUT,
+	CURPARSE_RUN
 };
 
 enum Error{
@@ -31,15 +33,18 @@ enum Error{
 	NoConfigEntryName,
 	TooLongConfigEntryName,
 	UnknownConfigurationEntry,
-	UnknownCompiler
+	UnknownCompiler,
+	UnknownRunValue
 };
 
 struct SMakeConfig{
 	int8_t compiler;
+	int8_t afterbuildrun;
 	char* files;
 	char* compilerflags;
 	char* assemblerflags;
 	char* linkerflags;
+	char* output;
 };
 
 char *filename = "build.smake";
@@ -64,6 +69,9 @@ void ErrExit(enum Error err){
 		case UnknownCompiler:
 			printf(ANSI_COLOR_RED "Unknown compiler!\n" ANSI_COLOR_RESET);
 			exit(1);
+		case UnknownRunValue:
+			printf(ANSI_COLOR_RED "Unknown value of RUN parameter!\n" ANSI_COLOR_RESET);
+			exit(1);
 	}
 }
 
@@ -77,11 +85,14 @@ void PrintConfig(struct SMakeConfig *config){
 }
 
 void GetDirName(char* filepath){
-	char t[2];
-	t[0] = "/";
-	t[1] = '\0';
-	char *laststash = strrchr(filepath, (char*)&t);
-	if (laststash != NULL) *laststash = '\0';
+	int32_t last_index = 0;
+	for (int32_t i = 0; i < strlen(filepath); i++) if (filepath[i] == '/') last_index = i;
+	if (last_index == 0) return;
+	filepath[last_index] = '\0';
+}
+
+char* GetFileName(char* path){
+    return strrchr(path, '/') + 1;
 }
 
 void HandleGetCWDError(void){
@@ -148,14 +159,16 @@ struct SMakeConfig *ParseSMakeConf(char *path){
 	fread(&buf, sizeof(char), len, fd);
 	fclose(fd);
 	char compiler[4] = {0};
+	char run[2] = {0};
 	char *files = (char*)calloc(sizeof(char), 16384);
 	char *compilerflags = (char*)calloc(sizeof(char), 512);
 	char *assemblerflags = (char*)calloc(sizeof(char), 512);
 	char *linkerflags = (char*)calloc(sizeof(char), 512);
+	char *output = (char*)calloc(sizeof(char), 128);
 	enum ParseMode currentmode = NULLVAL;
 	uint8_t parsing_str = 0;
 	char tmpbuf[33] = {0};
-	int16_t tmpbuf_offset = 0;
+	uint64_t tmpbuf_offset = 0;
 	char *bufptr = (char*)&tmpbuf;
 	for (uint64_t i = 0; i < len; i++){
 		if (buf[i] == '\n') continue;
@@ -167,6 +180,8 @@ struct SMakeConfig *ParseSMakeConf(char *path){
 			else if (strcmp(bufptr, "compilerflags") == 0) currentmode = CURPARSE_COMPILERFLAGS;
 			else if (strcmp(bufptr, "assemblerflags") == 0) currentmode = CURPARSE_ASSEMBLERFLAGS;
 			else if (strcmp(bufptr, "linkerflags") == 0) currentmode = CURPARSE_LINKERFLAGS;
+			else if (strcmp(bufptr, "run") == 0) currentmode = CURPARSE_RUN;
+			else if (strcmp(bufptr, "output") == 0) currentmode = CURPARSE_OUTPUT;
 			else ErrExit(UnknownConfigurationEntry);
 			continue;
 		}
@@ -189,6 +204,12 @@ struct SMakeConfig *ParseSMakeConf(char *path){
 					case CURPARSE_LINKERFLAGS:
 						bufptr = linkerflags;
 						break;
+					case CURPARSE_RUN:
+						bufptr = (char*)&run;
+						break;
+					case CURPARSE_OUTPUT:
+						bufptr = output;
+						break;
 				}
 			}
 			else if (parsing_str == 1){
@@ -206,23 +227,39 @@ struct SMakeConfig *ParseSMakeConf(char *path){
 	if (strcmp(compiler, "gcc") == 0) config->compiler = 0;
 	else if (strcmp(compiler, "g++") == 0) config->compiler = 1;
 	else ErrExit(UnknownCompiler);
+	if (run[0] == '0') config->afterbuildrun = 0;
+	else if (run[0] == '1') config->afterbuildrun = 1;
+	else ErrExit(UnknownRunValue);
 	config->files = files;
 	config->compilerflags = compilerflags;
 	config->assemblerflags = assemblerflags;
 	config->linkerflags = linkerflags;
+	config->output = output;
 	return config;
 }
 
-void ProcessFile(struct SMakeConfig* config, char* filepath){
-	char tmp[1024];
-	memcpy((char*)&tmp, filepath, 1024);
-	GetDirName((char*)&tmp);
-	printf((char*)&tmp);
+int ProcessFile(struct SMakeConfig* config, char* directory, char* src_filepath){
+	char parent_dir[1024];
+	char command[2048];
+	memcpy((char*)&parent_dir, directory, 1024);
+	GetDirName((char*)&parent_dir);
+	snprintf(command, sizeof(command), "mkdir -p %s", parent_dir);
+	int32_t retcode = system(command);
+	if (retcode != 0) return -1;
+	memset((char*)&command, 0x00, sizeof(command));
+	snprintf(command, sizeof(command), "cp %s %s", src_filepath, directory);
+	retcode = system(command);
+	if (retcode != 0) return -1;
+	memset((char*)&command, 0x00, sizeof(command));
+	char* filename = GetFileName(src_filepath);
+	snprintf(command, sizeof(command), "%s %s/%s %s", config->compiler == 0 ? "gcc" : "g++", directory, filename, config->compilerflags);
+	snprintf(((char*)&command) + strlen(command), sizeof(command) - strlen(command), " -o %s", config->output);
+	system(command);
 }
 
-void RunBuild(void){
+int RunBuild(void){
 	char *buf = (char*)calloc(sizeof(char), PATH_MAX + 64);
-	char *curpath;
+	char curpath[PATH_MAX];
 	errno = 0;
 	if (getcwd(curpath, PATH_MAX) == NULL) HandleGetCWDError();
 	CombinePaths(buf, curpath, filename);
@@ -236,25 +273,40 @@ void RunBuild(void){
 	snprintf(command_foldercreation, sizeof(command_foldercreation), "mkdir %s", directory);
 	system(command_foldercreation);
 	char *filepath = (char*)calloc(sizeof(char), 2048);
+	char *src_filepath = (char*)calloc(sizeof(char), 2048);
 	char parsed_filename[1024] = {0};
 	uint16_t offset = 0;
-	PrintConfig(config);
 	for (uint64_t i = 0; i < strlen(config->files); i++){
-		if (config->files[i] == ' ' && config->files[i] == '\0'){
-			CombinePaths(filepath, directory, parsed_filename);			
-			ProcessFile(config, parsed_filename);
+		if (config->files[i] == ' ' || config->files[i] == '\0'){
+			CombinePaths(filepath, directory, parsed_filename);
+			CombinePaths(src_filepath, buf, parsed_filename);
+			int32_t retcode = ProcessFile(config, filepath, src_filepath);
+			if (retcode != 0) return -1;
 			memset((char*)&parsed_filename, 0x00, sizeof(parsed_filename));
+			memset(filepath, 0x00, 2048);
+			memset(src_filepath, 0x00, 2048);
 			offset = 0;
 			continue;
 		}
 		parsed_filename[offset++] = config->files[i];
 	}
-	CombinePaths(filepath, directory, parsed_filename);			
-	ProcessFile(config, filepath);
+	CombinePaths(filepath, directory, parsed_filename);
+	CombinePaths(src_filepath, curpath, parsed_filename);
+	int retcode = ProcessFile(config, directory, src_filepath);
+	if (retcode != 0) return -1;
 	memset((char*)&parsed_filename, 0x00, sizeof(parsed_filename));
+	free(filepath);
+	free(src_filepath);
 	char removecommand[32] = {0};
 	snprintf(removecommand, sizeof(removecommand), "rm -rf %s", directory);
 	system(removecommand);
+	if (config->afterbuildrun == 1){
+		char buffer[256] = {0};
+		buffer[0] = '.';
+		buffer[1] = '/';
+		strcpy(((char*)&buffer) + 2, config->output);
+		system(buffer);
+	}
 }
 
 int main(int argc, char *argv[]){
